+++ .gitignore
@@ -1,1 +1,4 @@
 target/
+.project
+.project
+.settings/org.eclipse.wst.common.component
