+++ pom.xml
@@ -44,6 +44,11 @@
     	<artifactId>spring-web</artifactId>
     	<version>4.3.7.RELEASE</version>
     </dependency>
+    <dependency>
+    <groupId>org.springframework</groupId>
+    <artifactId>spring-orm</artifactId>
+    <version>4.3.7.RELEASE</version>
+</dependency>
         <dependency>
     	<groupId>mysql</groupId>
     	<artifactId>mysql-connector-java</artifactId>
