+++ src/main/java/edu/dmacc/books/application/BooksApplication.java
@@ -1,0 +1,13 @@
+package edu.dmacc.books.application;
+
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+import edu.dmacc.books.config.AppConfig;
+
+public class BooksApplication {
+
+	public static void main(String[] args) {
+		ApplicationContext context2 = new AnnotationConfigApplicationContext(AppConfig.class);
+	}
+}
