+++ src/main/java/edu/dmacc/books/config/AppConfig.java
@@ -1,0 +1,11 @@
+package edu.dmacc.books.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("edu.dmacc.books")
+public class AppConfig {
+	
+	
+}
