--- src/main/java/edu/dmacc/spring/books/Books.java
@@ -1,67 +1,0 @@
-package edu.dmacc.spring.books;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="books")
-public class Books {
-	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
-	private String name;
-	private String author;
-	private String detail;
-	private String copyrightyear;
-	private double isbn;
-	private String country;
-	private Borrower borrower;
-	
-	public int getId() {
-		return id;
-	}
-	public void setId(int id) {
-		this.id = id;
-	}
-	
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	public String getAuthor() {
-		return author;
-	}
-	public void setAuthor(String author) {
-		this.author = author;
-	}
-	public String getDetail() {
-		return detail;
-	}
-	public void setDetail(String detail) {
-		this.detail = detail;
-	}
-	public String getCopyrightyear() {
-		return copyrightyear;
-	}
-	public void setCopyrightyear(String copyrightyear) {
-		this.copyrightyear = copyrightyear;
-	}
-	public double getIsbn() {
-		return isbn;
-	}
-	public void setIsbn(double isbn) {
-		this.isbn = isbn;
-	}
-	public String getCountry() {
-		return country;
-	}
-	public void setCountry(String country) {
-		this.country = country;
-	}
-	}
