--- src/main/java/edu/dmacc/spring/books/Borrower.java
@@ -1,92 +1,0 @@
-package edu.dmacc.spring.books;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity 
-@Table(name="borrowers")
-public class Borrower {
-	@Id
-	private int borrowerID;
-	private String firstname;
-	private String lastname;
-	private String email;
-	private int phone1;
-	private int phone2;
-	private String address1;
-	private String address2;
-	private String city;
-	private String state;
-	private int zip;
-	public int getBorrowerID() {
-		return borrowerID;
-	}
-	public void setBorrowerID(int borrowerID) {
-		this.borrowerID = borrowerID;
-	}
-	public String getFirstname() {
-		return firstname;
-	}
-	public void setFirstname(String firstname) {
-		this.firstname = firstname;
-	}
-	public String getLastname() {
-		return lastname;
-	}
-	public void setLastname(String lastname) {
-		this.lastname = lastname;
-	}
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	public int getPhone1() {
-		return phone1;
-	}
-	public void setPhone1(int phone1) {
-		this.phone1 = phone1;
-	}
-	public int getPhone2() {
-		return phone2;
-	}
-	public void setPhone2(int phone2) {
-		this.phone2 = phone2;
-	}
-	public String getAddress1() {
-		return address1;
-	}
-	public void setAddress1(String address1) {
-		this.address1 = address1;
-	}
-	public String getAddress2() {
-		return address2;
-	}
-	public void setAddress2(String address2) {
-		this.address2 = address2;
-	}
-	public String getCity() {
-		return city;
-	}
-	public void setCity(String city) {
-		this.city = city;
-	}
-	public String getState() {
-		return state;
-	}
-	public void setState(String state) {
-		this.state = state;
-	}
-	public int getZip() {
-		return zip;
-	}
-	public void setZip(int zip) {
-		this.zip = zip;
-	}
-	
-	
-	
-
-}
