--- src/main/java/edu/dmacc/spring/books/config/WebConfig.java
@@ -1,36 +1,0 @@
-package edu.dmacc.spring.books.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.web.servlet.config.annotation.DefaultServletHandlerConfigurer;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-
-@EnableWebMvc
-@Configuration
-@ComponentScan("edu.dmacc.spring.books")
-public class WebConfig extends WebMvcConfigurerAdapter 
-{
-
-	    @Override
-	    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-	        registry.addResourceHandler("/resources/**").addResourceLocations("/resources/");
-	    }
-	 
-	    @Override
-	    public void configureDefaultServletHandling(DefaultServletHandlerConfigurer configurer) {
-	        configurer.enable();
-	    }
-	 
-	    @Bean
-	    public InternalResourceViewResolver jspViewResolver() {
-	        InternalResourceViewResolver bean = new InternalResourceViewResolver();
-	        bean.setPrefix("/WEB-INF/pages/");
-	        bean.setSuffix(".jsp");
-	        return bean;
-	    }
-}
