+++ src/main/webapp/WEB-INF/pages/SearchPages/BookSearchResult.jsp
