+++ src/main/webapp/WEB-INF/pages/mainForm.jsp
@@ -6,13 +6,18 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>Spring MVC Form Handling</title>
+<title>Main Menu</title>
 </head>
 <body>
 <h2>Books Menu</h2>
 	<ul>
-		<li><a href = "checkPage.mvc">Check Book In</a></li>
-		<li><a href = "viewAll.mvc">View All Users</a></li>
+		<li><a href = "BookSearch.mvc">Search and Borrow</a></li>
+		<li><a href = "CheckIn.mvc">Check Books Back In</a></li>
+		<li><a href = "CheckPages.mvc">Check Books Out</a></li>
+		<li><a href = "Books.mvc">Book Menu</a></li>
+		<li><a href = "BorrowerMenu.mvc">Borrowers Menu</a></li>
+		<li><a href = "ReportPage.mvc">Reports</a></li>
+        <!--  <li><a href = "viewAll.mvc">View All Users</a></li>  -->
 	</ul>
 </body>
 </html>
