--- src/main/webapp/WEB-INF/pages/maintainence.jsp
@@ -1,20 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
-    pageEncoding="ISO-8859-1"%>
-   <%@ taglib uri = "http://www.springframework.org/tags/form" prefix="mvc" %>
-   <%@ page isELIgnored="false" %>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>Maintenance</title>
-</head>
-<body>
-<h2>Maintenance</h2>
-	<ul>
-		<li><a href = "addBook.mvc">Add a Book</a></li>
-		<li><a href = "modBook.mvc">Modify a Book/a></li>
-		<li><a href = "delBook.mvc">Delete a Book/a></li>
-	</ul>
-</body>
-</body>
-</html>
