+++ src/main/webapp/WEB-INF/web.xml
@@ -13,7 +13,7 @@
     </init-param>
     <init-param>
         <param-name>contextConfigLocation</param-name>
-        <param-value>edu.dmacc.spring.books.config.WebConfig</param-value>
+        <param-value>edu.dmacc.books.config.AppConfig</param-value>
     </init-param>
   	<load-on-startup>1</load-on-startup>
   </servlet>
